Migrate to Modern Libs
Library upgrades:
- SDL 1.2                -> SDL3 (>=3.2.0)
- SDL_mixer 1.2           -> SDL3_mixer (>=3.0.0) - full API redesign
                             (Mix_Chunk/channels -> MIX_Track/MIX_Audio/MIX_Mixer)
- SDL_image 1.2           -> SDL3_image (>=3.0.0)
- SDL_ttf 1.2             -> SDL3_ttf (>=3.0.0)
- Lua 5.0/5.1             -> Lua >=5.5
- FriBidi                 unchanged (still optional)
- Boost                   unchanged (still optional, --with-boost -> WITH_BOOST)

Removed:
- Bundled SDL_gfx (unmaintained) - replaced its 2 call sites with a
  local DrawPrimitives helper
- SMPEG intro-movie support - no SDL3 port exists, unmaintained since ~2004
- X11 dependency - was only used for a caption workaround SDL3 handles
  natively

Build system:
- automake/autoconf -> CMake (>=3.16)
- Project renamed fillets-ng -> fillets, version bumped 1.0.1 -> 1.9

+++ src/gengine/MouseStroke.cpp
@@ -15,7 +15,7 @@
  * Create new stroke from MouseButtonEvent.
  */
 MouseStroke::MouseStroke(const SDL_MouseButtonEvent &event)
-    : m_loc(event.x, event.y)
+    : m_loc(static_cast<int>(event.x), static_cast<int>(event.y))
 {
     m_button = event.button;
 }
